+++ build_runner/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.7.12-prerelease
 
 - Build actions using `findAssets` will be more smartly invalidated.
+- Added a warning if using `serve` mode but no directories were found to serve.
 
 ## 0.7.11+1