+++ build_runner/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 - Build actions using `findAssets` will be more smartly invalidated.
 - Added a warning if using `serve` mode but no directories were found to serve.
+- The `--output` option now only outputs files that were required for the latest
+  build. Previously when switching js compilers you could end up with ddc
+  modules in your dart2js output, even though they weren't required. See
+  https://github.com/dart-lang/build/issues/1033.
+- The experimental `create_merged_dir` binary is now removed, it can't be easily
+  supported any more and has been replaced by the `--output` option.
 
 ## 0.7.11+1